Save and Reset Gain+Offset on Image capture

* save and reset gain on capture

* add gain to image metadata

* display gain from camera equipment by default instead of empty value

* fix nikon saving gain, add preselected and default gain instead of empty value, change gain to int where necessary

* Merge remote-tracking branch 'upstream/develop' into feature/gainreset

* Merge branch 'main' into feature/gainreset

* change short gain to int gain

* when camera is not connected and gain is displayed display (Camera) instead of DependencyProperty.UnsetValue


Former-commit-id: 6049e0d0e8597e6faf903b69dfb1d56b6a04045a
Former-commit-id: eca1c5e7cbf7752ff41bf56460aa7adb99db20e5

+++ NINA/Utility/Mediator/CameraMediator.cs
@@ -54,7 +54,7 @@ namespace NINA.Utility.Mediator {
             handler.SetBinning(x, y);
         }
 
-        public void SetGain(short gain) {
+        public void SetGain(int gain) {
             handler.SetGain(gain);
         }
